requirements.txt
kafka-python==2.1.2
requests==2.32.3
dotenv==0.9.9